combine ABFcore and ABF
